Added to and fixed proprietary-files.txt
I melded the proprietary-files.txt from this device tree with garwynn's, and then added/removed and corrected the path of certain files, according to the files present on my 16gb sprint s5. This list runs runs without any errors. I can't gaurantee it includes everything needed, but it does include every file from the two klte devices trees on git that were present on my device, and their correct paths. All files that are commented out are not present anywhere in the system folder of my device. I also added a few that were not included, and fixed the number versions of some files (ie libmmcamera_cac2_lib.so -> libmmcamera_cac3_lib.so).